site/requirements.txt
autopep8==1.6.0
certifi>=2022.12.07
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.6
Flask==2.2.2
PyJWT==2.6.0
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.2.2

mysql-connector-python~=8.0.32
bitcoinaddress~=0.1.5